app/pyproject.toml
[tool.poetry]
name = "construct"
version = "0.1.0"
description = "construction scheduling + pddl + llm agent"
authors = ["Nicholas Woodward <nicholas.ean.woodward@gmail.com>"]
license = "MIT"
packages = [{ include = "construct" }]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0"
langchain = "^0.0.315"
sqlalchemy = "^2.0"
typer = "^0.7.0"
pandas = "^2.0"  # for excel ingestion
openpyxl = "^3.1"  # for xlsx engine
openai = "0.27.8"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.3.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
construct = "construct.main:main"